ai-cli-assistant/requirements.txt
google-genai 
python-dotenv